SD playback status issues
When an M23 is issued to start an SD print, some hosts (like OctoPrint)
ask for build status (M27) pretty quickly.  Sailfish sets the build
running state only after it receives a Build start notification (153)
which is in the x3g file, but may not be the first instruction. So we
need a way to delay giving back the status until that happens. This
commit attempts to return "wait" for up to 3 seconds for the build to
start. Need to think of a more robust mechanism since the 153 could be
after another long wait for temperature?