Improve cancel
This commit applies to GPX being used as a submodule for OctoPrint-GPX.
Cancel can be called asynchronously with the main gcode processing loop
and so a few gcode commands can come through after the cancel and since
we do an immediate stop and reset, the coordinate system has been lost
so those enqueuing moves could be very random. This commit introduces a
cancelPending mode that eats commands until the gcode stream shows it is
in sync with the asynchronous cancel with (@clear_cancel) macro.