Assume M132 recalls 0 for A and B
Problem is that a fair amount of start gcode out there doesn't ever set
the other extruder to a known location so for single extruder prints we
run as if we don't know where we are (even though we do for all
practical purposes).  More correct start gcode would be:
G92 X0 Y0 Z0 A0 B0
M132 X Y Z

It sets X Y Z to 0 and then stomps on them with M132 because G92 A0 B0
will set X Y and Z anyway, but to whatever is in "current" in GPX
because x3g doesn't have a way to define position other than all 5
dimesions.