Let the firmware limit the maximum bed temperature
This relaxes the fix for issue #3 even further by removing the internal
GPX limit on HBP temperature.  Bit width of the parameter in the
protocol is 16-bit, so now will allow any unsigned value to be sent
over the protocol.

The only reference left is in macro parsing.  Temperatures <= 130 will
still be considered bed temperatures and > 130 nozzle temperatures.