short/long retry buffer overflow
Introduce a two-tier retry strategy for buffer overflow.  For a while,
retry sending on a short pause.  After that fails, go to a long pause
retry.  The intent of this is to keep the buffer full when executing
lots of short movements, but not hard spin when the queue is full of
long moves.