Add machine configuration flag for ABP
M106/M107 turn on/off the cooling fan on most machines. On a machine
with an Automated Build Platform (ABP) it operates the ABP. Prior GPX
always assumed that machines before the Replicator had an ABP, this
change adds two new machine types with ABP and M106/M107 operates the
cooling fan on all others.